Configure for heroku deployment
These changes allow for the following:

- configure heroku to utilize yarn v1 building
- configure the disord bot service to run as a worker dyno
- allow the hasura endpoints to be configured in an environment
- utilize ts-node in prod

Remaining Todos

**cleanly delineate devDependencies**

This might mean using the tsc compilation process in prod and using
node instead of ts-node, since we want to be able to prune extraneous
dev dependencies before the dyno starts up. This is configured by
`NODE_ENV` variable set in heroku, which is currently set to staging,
and defaults to production, which will prune devDeps after the build
step.